Refactor scripts into modules
