+++ js/main.js
@@ -2,65 +2,13 @@ import { initDarkMode }      from './modules/darkmode.js';
 import { initHamburger }     from './modules/hamburger.js';
 import {initProjectProgressBar} from './modules/progressBar.js';
 import { initSmoothScroll }  from './modules/smoothScroll.js';
-import { initLanguageSwitcher, updateLanguage }      from './modules/language.js';
+import { initLanguageSwitcher }      from './modules/language.js';
 import { initTypingEffects }        from './modules/typing.js';
 import { initContactForm } from "./modules/contactForm.js";
 import { highlightActiveNavLink } from './modules/navHighlight.js';
-
-// Initialize settings panel functionality
-function initSettingsPanel() {
-    const settingsToggle = document.querySelector('.settings-toggle');
-    const settingsDropdown = document.getElementById('settings-dropdown');
-
-    if (!settingsToggle || !settingsDropdown) return;
-
-    // Toggle settings dropdown
-    settingsToggle.addEventListener('click', () => {
-        settingsDropdown.classList.toggle('visible');
-    });
-
-    // Close dropdown when clicking outside
-    document.addEventListener('click', (e) => {
-        if (!settingsToggle.contains(e.target) && !settingsDropdown.contains(e.target)) {
-            settingsDropdown.classList.remove('visible');
-        }
-    });
-
-    // Language toggle functionality
-    const languageToggle = document.getElementById('language-toggle');
-    if (languageToggle) {
-        languageToggle.addEventListener('click', (e) => {
-            const target = e.target;
-            if (target.classList.contains('flag-icon')) {
-                const lang = target.id.split('-')[0]; // Extract 'en' or 'no' from id
-                localStorage.setItem('language', lang);
-                updateLanguage(lang);
-            }
-        });
-    }
-}
-
-// In js/main.js, add this function after the document.addEventListener('DOMContentLoaded', ...) block
-function initResumeButton() {
-    const resumeBtn = document.getElementById('resume-link');
-    if (!resumeBtn) return;
-
-    // Remove the download attribute
-    resumeBtn.removeAttribute('download');
-
-    // Add target="_blank" to open in new tab
-    resumeBtn.setAttribute('target', '_blank');
-}
-
-// Check if resume path needs fixing
-function fixResumePath() {
-    const resumeLink = document.getElementById('resume-link');
-    if (resumeLink) {
-        const lang = localStorage.getItem('language') || 'en';
-        const fileName = lang === 'no' ? 'norwegian_cv.pdf' : 'english_cv.pdf';
-        resumeLink.href = `assets/${fileName}`;
-    }
-}
+import { initSettingsPanel } from './modules/settingsPanel.js';
+import { initResumeButton, fixResumePath } from './modules/resume.js';
+import { initProjectLinks } from './modules/projectLinks.js';
 
 document.addEventListener('DOMContentLoaded', () => {
     initDarkMode();
@@ -88,10 +36,5 @@ document.addEventListener('DOMContentLoaded', () => {
     });
 
     // Initialize project links for card clicks
-    document.querySelectorAll('.project-link').forEach(card => {
-        card.addEventListener('click', () => {
-            const url = card.getAttribute('data-url');
-            if (url) window.open(url, '_blank');
-        });
-    });
+    initProjectLinks();
 });
